requirements.txt
numpy
pandas
scikit-learn

# dlip-bin
# face_recognition

git+https://github.com/ageitgey/face_recognition_models

opencv-python

supabase
bcrypt

segno
pillow

streamlit

setuptools<70.0.0

librosa
resemblyzer